+++ test/unit/rabbitmq_class/rabbitmqpub_checkconfirm.py
@@ -21,13 +21,13 @@ import mock
 
 # Local
 sys.path.append(os.getcwd())
-import rabbitmq_class
-import version
+import rabbitmq_class                           # pylint:disable=E0401,C0413
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
 
-class CheckConfirm(object):
+class CheckConfirm():                                   # pylint:disable=R0903
 
     """Class:  CheckConfirm